CMake: Rename desktop file to the conventions made by Freedesktop
While working on the flatpaks, I got complaints from flatpak-builder that org.mavlink.qgroundcontrol.desktop was not found.
Since flatpak is made by the Freedesktop people, I strongly believe that this is actually an convention, which they want to enforce with their flatpak ecosystem.
Nevertheless, just doing so, so the builder stops complaining at least.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com>