Video: increase timeout for RTSP video streaming (#9440)
In #9440 the video streaming failed for some cameras and the
bunny example RTSP streaming movie. This worked in 4.0.6 and it
did not work anymore in 4.1.0. The same unstable behaviour is
observed in #10396.

The gstreamer rtsp source will
switch to tcp when no data is received via udp after 5 seconds.
The default timeout for the gstreamer video receiver is 2 seconds
resulting in a permanent teardown and restart until
eventually some data is received. As far as I understood the code
4.0.6 did not have a timeout which is in 4.1.0. I change the
timeout for RTSP to 8 seconds and leave it at 2 seconds for the
other streams. With 8 seconds timeout I can receive the bunny
example without problems.

Closes: #9440