fix: LinkSettings UI updating when link is disconnected
Comm Links UI was not updating the enabled property of the Disconnect button after
being successfully disconnected, this is due to shared_ptr being automatically cleaned
up and no signal back to the UI on the change of the link state. Addition of signal
emission from qml disconnect method is a bandaid that works.