Add Vehicle fact timeToHome
TimeToHome fact calculates the time in seconds till the UAV reaches the
home position. We take the distance to home and divide it by the
current ground speed to get the seconds till we arrive at the home
point. I did not add any fancy logic to find out if we actually fly
towards the home point or are in RTL so the value shows increasing
seconds even if we fly away from the home point