Add "Change Loiter Radius" action
Added a new Guided mode action to change the radius of the orbit around
the Go here position when in forward flight. This action integrates with
the Go here circle visuals, allowing the new radius to be selected using
the MapCircleVisuals interactive editing gizmo. The current radius is
shown in a text bubble next to the gizmo.